feat: 파일 분리 및 함수 기능 추가
CoffeeShop, Person로 파일 분리 및 함수 일부 수정
